third_party/WebKit/LayoutTests/crypto/generateKey.html

<!DOCTYPE html>
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests cypto.subtle.generateKey.");

jsTestIsAsync = true;

// Note that fractional numbers are truncated, so this length should be
// interpreted as 1024.
aesCbc = { name: 'aes-cbc', length: 1024.9 };
extractable = true;
keyUsages = ['encrypt', 'decrypt'];

// length property is missing.
invalidAesKeyGen = { name: 'aes-cbc' };
shouldThrow("crypto.subtle.generateKey(invalidAesKeyGen, extractable, keyUsages)");

// length is invalid (outside of range of "unsigned short")
invalidAesKeyGen = { name: 'aes-cbc', length: 70000 };
shouldThrow("crypto.subtle.generateKey(invalidAesKeyGen, extractable, keyUsages)");

// length is invalid (outside of range of "unsigned short")
invalidAesKeyGen = { name: 'aes-cbc', length: -3 };
shouldThrow("crypto.subtle.generateKey(invalidAesKeyGen, extractable, keyUsages)");

// keyUsages is invalid.
shouldThrow("crypto.subtle.generateKey(aesCbc, extractable, -1)");

// Invalid length
invalidHmac256 = { name: 'hmac', hash: {name: 'sha-256' }, length:-3 };
shouldThrow("crypto.subtle.generateKey(invalidHmac256, false, ['sign'])");

hmacSha256 = { name: 'hmac', hash: {name: 'sha-256' } };
hmacSha256b = { name: 'hmac', hash: {name: 'sha-256' }, length:48 };

var promise0 = crypto.subtle.generateKey(aesCbc, extractable, keyUsages);
var promise1 = crypto.subtle.generateKey(hmacSha256, false, ['sign']);
var promise2 = crypto.subtle.generateKey(hmacSha256b, false, ['sign']);

Promise.every(promise0, promise1, promise2).then(function(results)
{
    key = results[0];
    shouldBe("key.type", "'private'")
    shouldBe("key.extractable", "true")
    shouldBe("key.algorithm.name", "'AES-CBC'")
    shouldBe("key.algorithm.length", "1024")
    shouldBe("key.usages.join(',')", "'encrypt,decrypt'")

    key = results[1];
    shouldBe("key.type", "'private'")
    shouldBe("key.extractable", "false")
    shouldBe("key.algorithm.name", "'HMAC'")
    shouldBe("key.algorithm.hash.name", "'SHA-256'")
    shouldBe("key.algorithm.length", "null")
    shouldBe("key.usages.join(',')", "'sign'")

    key = results[2];
    shouldBe("key.type", "'private'")
    shouldBe("key.extractable", "false")
    shouldBe("key.algorithm.name", "'HMAC'")
    shouldBe("key.algorithm.hash.name", "'SHA-256'")
    shouldBe("key.algorithm.length", "48")
    shouldBe("key.usages.join(',')", "'sign'")

    finishJSTest();
});

</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
